Refactor payment package: rename service package and update handler imports
