+++ payment/internal/handlers/payment-handler.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"e-commerce-backend/payment/dbs"
-	"e-commerce-backend/payment/internal/services/payment"
+	"e-commerce-backend/payment/internal/services"
 	"e-commerce-backend/shared/middlewares"
 	"net/http"
 
@@ -10,7 +10,7 @@ import (
 )
 
 func PaymentHandler(r *mux.Router) {
-	paymentService := payment.NewPaymentService(dbs.DB)
+	paymentService := services.NewPaymentService(dbs.DB)
 
 	r.Handle("/user/{id}/payment", middlewares.AuthMiddleware(http.HandlerFunc(paymentService.GetPayment))).Methods("POST")
 	r.Handle("/user/{id}/payment/add", middlewares.AuthMiddleware(http.HandlerFunc(paymentService.AddToPayment))).Methods("POST")
